+++ libraries/AtomicWeight/AtomicWeight.h
@@ -3,14 +3,19 @@
 //    FILE: AtomicWeight.h
 //  AUTHOR: Rob Tillaart
 //    DATE: 2022-03-09
-// VERSION: 0.1.4
+// VERSION: 0.1.5
 // PURPOSE: Arduino library for atomic weights
 //     URL: https://github.com/RobTillaart/AtomicWeight
 
 
 #include "Arduino.h"
 
-#define ATOMIC_WEIGHT_LIB_VERSION         (F("0.1.4"))
+#define ATOMIC_WEIGHT_LIB_VERSION         (F("0.1.5"))
+
+
+#ifndef ATOMIC_WEIGHT_MAX_SPLIT_LIST
+#define ATOMIC_WEIGHT_MAX_SPLIT_LIST 20
+#endif
 
 
 /////////////////////////////////////////////////////////////////////////
@@ -33,13 +38,18 @@ public:
   float   weight(const uint8_t el);
   //  if (el != NULL) weights one element in a formula, e.g el == "H"
   //  if (el == NULL) weights the whole formula
-  float   weight(const char * formula, const char * el = NULL);
+  float   weight(const char * formula, const char * abbrev = NULL);
   //  mass percentage of one element in a formula.
-  float   massPercentage(const char * formula, const char * el);
+  float   massPercentage(const char * formula, const char * abbrev);
 
 
   char *  name(const uint8_t el);
   uint8_t find(const char * abbrev);
+
+
+  //  CONVERSION
+  float   moles2grams(const char * formula, float moles = 1.0);
+  float   grams2moles(const char * formula, float grams = 1.0);
 
 
   //  DEBUG
@@ -67,12 +77,12 @@ private:
   const float _weightFactor = 1.0 / 222.909;
 
   //  if (el == NULL) ==> whole weight otherwise only of element.
-  float       _weight(char sep, const char * el);
-  uint32_t    _count(const char sep, const char * el);
+  float       _weight(char sep, const char * abbrev);
+  uint32_t    _count(const char sep, const char * abbrev);
   char        *p;  //  for _weight() and _count()
 
   //  for splitElements
-  uint8_t     _elems[20];  //  max 20 elements in formula.
+  uint8_t     _splitList[ATOMIC_WEIGHT_MAX_SPLIT_LIST]; //  default 20
   uint8_t     _found;
 };